requirements.txt
Django==5.2.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1